backend/requirements.txt
requests==2.26.0
Django==3.2.3
djangorestframework==3.12.4
django-filter==22.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
djoser==2.1.0
python-decouple==3.5
drf-extra-fields==3.4.1
Pillow==9.5.0
isort==5.11.5
flake8==6.0.0
python-dotenv==1.0.0

